+++ test/session_test.rb
@@ -113,27 +113,33 @@ class SessionTest < ActiveSupport::TestCase
       assert_equal user, session.user
       assert_equal user.id, session.user_id
       assert_equal true, session.persist?
+      assert_equal true, session.save
       
       User.unstub(:find_for_session)
     end
     
     should "validate with an api key" do
+      Challah.options[:api_key_enabled] = true
+      
       user = Factory(:user, :api_key => '123456abcdefg')
       
       User.stubs(:find_for_session).returns(user)
       
       session = Session.new
       session.ip = '127.0.0.1'
-      session.api_key = '123456abcdefg'
+      session.key = '123456abcdefg'
       
       assert_equal true, session.valid?
       assert_equal user, session.user
       assert_equal user.id, session.user_id      
       assert_equal false, session.persist?
+      assert_equal false, session.save
       
       user.expects(:successful_authentication!).with('127.0.0.1').once
       
       User.unstub(:find_for_session)
+      
+      Challah.options[:api_key_enabled] = false
     end 
     
     should "reject if password is incorrect" do